PRSD-362: Rename local/api/controller to fit convention
